+++ src/FileBrowserGUI.cpp
@@ -40,7 +40,7 @@ namespace ALZ {
 		ImGui::SameLine();
 		ImGui::TextColored(ImVec4(0.0f, 1.0f, 0.0f, 1.0f), m_CurrentselectedFilePath.c_str());
 		ImGui::PushItemWidth(-1);
-		ImGui::ListBoxHeader("##empty", 0, std::distance(fs::directory_iterator(m_CurrentFolder), fs::directory_iterator{}) + 5);
+		ImGui::ListBoxHeader("##empty", 0, (int)std::distance(fs::directory_iterator(m_CurrentFolder), fs::directory_iterator{}) + 5);
 
 		//check if we can go back
 		if (std::count(m_CurrentFolder.begin(), m_CurrentFolder.end(), '\\') > 0) {
